+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- `Client.get_collection` for static catalogs [#782](https://github.com/stac-utils/pystac-client/pull/782)
+
 ## [v0.8.5] - 2024-10-23
 
 ### Fixed